Rename the project to Cartographer
This fork has been created to add features that aren't necessarily
in the scope of the simplicity vision of the parent project. In the
future, this will include Celery for task running, a system for
scheduling recurring scans, etc. This rename serves to disambiguate
the fork as our aim is different than that of rainmap-lite. Much
credit is owed to the project for the foundation.